converted light to percentage

+++ embedded/src/data_flow/data_flow.h
@@ -14,7 +14,7 @@ typedef struct {
 // Payload Data Struct for exchanging data between Core0 and Core1
 typedef struct {
     volatile uint64_t time_stamp;
-    volatile uint16_t ADC_Data; // this only has 12 bits of precision, we lose 4 bits
+    volatile float ADC_Data; // this only has 12 bits of precision, we lose 4 bits
     volatile DHT20_Reading DHT20_Data;   //  store temp & humidity sensor data
     volatile bool DHT20_Data_Valid;     
 } Payload_Data;
